smart-parking-k3d/services/aggregator/requirements.txt
paho-mqtt==1.6.1
kubernetes==28.1.0
